requirements.txt
Flask-MySQLdb==0.2.0
SQLAlchemy==1.4.22
passlib==1.7.4
Flask==2.0.1
gunicorn==20.1.0
mysql-connector-python==8.0.23
hashlib==20081119